action table in one template
